requirements.txt
CherryPy==3.2.2
Django==1.4.3
FormEncode==1.2.6
Genshi==0.6
South==0.7.6
argparse==1.2.1
configobj==4.7.2
distribute==0.6.24
dj-database-url==0.2.1
django-breadcrumbs==1.1.3-p1
django-crispy-forms==1.2.3
django-registration==0.8
django-social-auth==0.7.18
django-tinymce==1.5.1b4
django-visitor==0.1.7
httplib2==0.7.7
oauth2==1.5.211
psycopg2==2.4.6
python-openid==2.2.5
wsgiref==0.1.2
